Update code
This is the code for the base scenarios. It's modular, so if you change the variables, it should change everything else.

The payoff matrix is a 3d array:
[[[0, 0]
   [0, 0,]]
  [[0, 0]
   [0, 0,]]]
it has two 2d arrays, each representing a 2x2 payoff matrix for each player:
[[0, 0]
 [0, 0]]
so if the 3d array was:
[[[4, 2]
   [-2, 0,]]
  [[0, -2]
   [2, 4,]]]
then you could write it like this:
____________________
|    4, 0    | 2, -2    |
|___________________
|    -2, 2    | 0, 4    |
____________________

All of the variables we had discussed (such as disagreement incurring a -2 penalty or favor giving +3) are global variables.
BaseMatrix is also global, but all of the functions return a result, rather than modifying this matrix. So if you wanted to print the anti-god scenario, you would do:
print(AntiGod(BaseMatrix))
which would then take the base matrix (which is a 2x2x2 filled with 0's), and apply the Favor, Disfavor, Agreement and Disagreement factors (using the antiMirth, antiWrath, coexistence and strife functions respectively).

All of the comments give the name, input types if if it's a function, and an explanation of what it does. I kept the names biblical but simple. Let me know what you think!
- Brandon